requirements.txt
Streamlit>=0.71.0
python-i18n[YAML]>=0.3.9
